fake_pub_sub_on_gcp/requirements_consumer.txt
google-cloud-pubsub
pandas
google-cloud-aiplatform
pyarrow